requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extra-views==0.13.0
djangorestframework==3.12.2
pkg-resources==0.0.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
